requirements.txt
fastapi==0.115.7
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pandas==2.2.2
SQLAlchemy==2.0.31
pytest==8.1.1
httpx==0.28.1
sqlmodel==0.0.22
alembic==1.14.1
validate-docbr==1.10.0
psycopg2==2.9.10
uvicorn
python-dotenv
